+++ examples/scripts/search/conjunctions/create_conjunction_search_object.py
@@ -1,5 +1,6 @@
 import datetime
 import pyaurorax
+import pprint
 
 
 def main():
@@ -9,10 +10,9 @@ def main():
     # set up params
     start = datetime.datetime(2020, 1, 1, 0, 0, 0)
     end = datetime.datetime(2020, 1, 1, 6, 59, 59)
-    ground_params = [{"programs": ["themis-asi"]}]
-    space_params = [{"programs": ["swarm"]}]
-    events_params = [{"programs": ["events"]}]
     distance = 200
+    ground_params = [aurorax.search.GroundCriteriaBlock(programs=["themis-asi"])]
+    space_params = [aurorax.search.SpaceCriteriaBlock(programs=["swarm"])]
 
     # create search object
     s = pyaurorax.search.ConjunctionSearch(
@@ -21,12 +21,15 @@ def main():
         end,
         ground=ground_params,
         space=space_params,
-        events=events_params,
         distance=distance,
     )
 
     print()
     print(s)
+    print()
+    s.pretty_print()
+    print()
+    pprint.pprint(s.query)
     print()