+++ examples/scripts/search/conjunctions/describe_conjunction_search.py
@@ -9,10 +9,9 @@ def main():
     # set up params
     start = datetime.datetime(2020, 1, 1, 0, 0, 0)
     end = datetime.datetime(2020, 1, 1, 6, 59, 59)
-    ground_params = [{"programs": ["themis-asi"]}]
-    space_params = [{"programs": ["swarm"]}]
-    events_params = [{"programs": ["events"]}]
     distance = 200
+    ground_params = [aurorax.search.GroundCriteriaBlock(programs=["themis-asi"])]
+    space_params = [aurorax.search.SpaceCriteriaBlock(programs=["swarm"])]
 
     # create search object
     s = pyaurorax.search.ConjunctionSearch(
@@ -21,7 +20,6 @@ def main():
         end,
         ground=ground_params,
         space=space_params,
-        events=events_params,
         distance=distance,
     )