+++ examples/scripts/search/conjunctions/legacy/describe_conjunction_search.py
@@ -9,23 +9,25 @@ def main():
     # set up params
     start = datetime.datetime(2020, 1, 1, 0, 0, 0)
     end = datetime.datetime(2020, 1, 1, 6, 59, 59)
+    ground_params = [{"programs": ["themis-asi"]}]
+    space_params = [{"programs": ["swarm"]}]
+    events_params = [{"programs": ["events"]}]
     distance = 200
-    ground_params = [aurorax.search.GroundCriteriaBlock(programs=["themis-asi"])]
-    space_params = [aurorax.search.SpaceCriteriaBlock(programs=["swarm"])]
 
     # create search object
-    s = aurorax.search.conjunctions.search(
+    s = pyaurorax.search.ConjunctionSearch(
+        aurorax,
         start,
         end,
         ground=ground_params,
         space=space_params,
+        events=events_params,
         distance=distance,
-        verbose=True,
     )
 
-    # create the custom import file
-    filename = aurorax.search.conjunctions.swarmaurora.create_custom_import_file(s)
-    print("\nSwarm-Aurora custom import file: %s\n" % (filename))
+    print()
+    print(aurorax.search.conjunctions.describe(search_obj=s))
+    print()
 
 
 # ----------