+++ examples/scripts/search/conjunctions/legacy/get_conjunction_search_request_status.py
@@ -1,4 +1,5 @@
 import datetime
+import time
 import pprint
 import pyaurorax
 
@@ -10,9 +11,9 @@ def main():
     # set up params
     start = datetime.datetime(2020, 1, 1, 0, 0, 0)
     end = datetime.datetime(2020, 1, 1, 6, 59, 59)
+    ground_params = [{"programs": ["themis-asi"]}]
+    space_params = [{"programs": ["swarm"]}]
     distance = 200
-    ground_params = [aurorax.search.GroundCriteriaBlock(programs=["themis-asi"])]
-    space_params = [aurorax.search.SpaceCriteriaBlock(programs=["swarm"])]
 
     # create search object
     s = aurorax.search.conjunctions.search(
@@ -23,19 +24,22 @@ def main():
         distance=distance,
         return_immediately=True,
     )
+    print(s)
 
-    # wait for data
-    print("\nWaiting for request to complete ...")
-    s.wait()
+    # get status
+    print("Getting request status ...")
+    s.update_status()
+    pprint.pprint(s.status)
+    print("----------------------------\n")
 
-    # get data
-    print("Get data ...")
-    s.get_data()
+    # if the request isn't done, wait continuously
+    print("Waiting for request to complete ...")
+    while (s.completed is False):
+        time.sleep(1)
+        s.update_status()
 
-    # print data
-    print()
-    pprint.pprint(s.data)
-    print()
+    # print status
+    pprint.pprint(s)
 
 
 # ----------