+++ examples/scripts/search/conjunctions/legacy/search_conjunction_synchronous.py
@@ -1,3 +1,4 @@
+import pprint
 import datetime
 import pyaurorax
 
@@ -9,11 +10,11 @@ def main():
     # set up params
     start = datetime.datetime(2020, 1, 1, 0, 0, 0)
     end = datetime.datetime(2020, 1, 1, 6, 59, 59)
+    ground_params = [{"programs": ["themis-asi"]}]
+    space_params = [{"programs": ["swarm"]}]
     distance = 200
-    ground_params = [aurorax.search.GroundCriteriaBlock(programs=["themis-asi"])]
-    space_params = [aurorax.search.SpaceCriteriaBlock(programs=["swarm"])]
 
-    # create search object
+    # perform search
     s = aurorax.search.conjunctions.search(
         start,
         end,
@@ -23,9 +24,12 @@ def main():
         verbose=True,
     )
 
-    # create the custom import file
-    filename = aurorax.search.conjunctions.swarmaurora.create_custom_import_file(s)
-    print("\nSwarm-Aurora custom import file: %s\n" % (filename))
+    # print data
+    print()
+    pprint.pprint(s.data)
+    print()
+
+    pprint.pprint(s.__dict__)
 
 
 # ----------