+++ examples/scripts/search/conjunctions/search_conjunction_synchronous.py
@@ -10,9 +10,9 @@ def main():
     # set up params
     start = datetime.datetime(2020, 1, 1, 0, 0, 0)
     end = datetime.datetime(2020, 1, 1, 6, 59, 59)
-    ground_params = [{"programs": ["themis-asi"]}]
-    space_params = [{"programs": ["swarm"]}]
     distance = 200
+    ground_params = [aurorax.search.GroundCriteriaBlock(programs=["themis-asi"])]
+    space_params = [aurorax.search.SpaceCriteriaBlock(programs=["swarm"])]
 
     # perform search
     s = aurorax.search.conjunctions.search(