+++ examples/scripts/search/conjunctions/search_conjunction_synchronous_with_events.py
@@ -1,3 +1,4 @@
+import pprint
 import datetime
 import pyaurorax
 
@@ -8,24 +9,29 @@ def main():
 
     # set up params
     start = datetime.datetime(2020, 1, 1, 0, 0, 0)
-    end = datetime.datetime(2020, 1, 1, 6, 59, 59)
+    end = datetime.datetime(2020, 2, 28, 23, 59, 59)
     distance = 200
     ground_params = [aurorax.search.GroundCriteriaBlock(programs=["themis-asi"])]
     space_params = [aurorax.search.SpaceCriteriaBlock(programs=["swarm"])]
+    events_params = [aurorax.search.EventCriteriaBlock()]
 
-    # create search object
+    # perform search
     s = aurorax.search.conjunctions.search(
         start,
         end,
         ground=ground_params,
         space=space_params,
         distance=distance,
+        events=events_params,
         verbose=True,
     )
 
-    # create the custom import file
-    filename = aurorax.search.conjunctions.swarmaurora.create_custom_import_file(s)
-    print("\nSwarm-Aurora custom import file: %s\n" % (filename))
+    # print data
+    print()
+    pprint.pprint(s.data)
+    print()
+
+    pprint.pprint(s.__dict__)
 
 
 # ----------