+++ examples/scripts/search/conjunctions/swarmaurora_open_in_browser.py
@@ -9,9 +9,9 @@ def main():
     # set up params
     start = datetime.datetime(2020, 1, 1, 0, 0, 0)
     end = datetime.datetime(2020, 1, 1, 6, 59, 59)
-    ground_params = [{"programs": ["themis-asi"]}]
-    space_params = [{"programs": ["swarm"]}]
     distance = 200
+    ground_params = [aurorax.search.GroundCriteriaBlock(programs=["themis-asi"])]
+    space_params = [aurorax.search.SpaceCriteriaBlock(programs=["swarm"])]
 
     # create search object
     s = aurorax.search.conjunctions.search(