+++ RELEASE_NOTES.md
@@ -1,6 +1,11 @@
 Version 1.12.0
 -------------------
 - Type hinting updates
+- Updates to conjunction, ephemeris, and data product searching in the `aurorax.search` module
+  - added new criteria block classes to help with searching: `GroundCriteriaBlock`, `SpaceCriteriaBlock`, `EventCriteriaBlock` (#54, #51)
+  - added new `MetadataFilterExpression` class to help with searching (#28)
+  - added support for custom lat/lon searching (#81)
+  - deprecated ground/space/events parameters as `Dict` type, included warnings
 
 
 Version 1.11.0